Add array ensure capacity with over-provisioning to core lib Continue implementation of mesh loading and saving

+++ lib/core/include/core/handle.h
@@ -1,11 +1,12 @@
 #ifndef CORE_HANDLE_H
 #define CORE_HANDLE_H
-#include <stddef.h>
+
+#include "defines.h"
 
 typedef struct {
-    size_t index;
+    u64 index;
 } handle_t;
 
-handle_t handle_create(size_t index);
+handle_t handle_create(u64 index);
 
 #endif
