+++ lib/core/src/handle.c
@@ -1,6 +1,6 @@
 #include "core/handle.h"
 
-handle_t handle_create(size_t index) {
+handle_t handle_create(u64 index) {
     handle_t handle;
     handle.index = index;
     return handle;
